[DDW-1028] Simplify block feature ui setup
