+++ source/features/blocks/ui/BlockSummary.scss
